+++ src/main/java/ic3/analyticsops/restapi/reply/print/AORestApiPrintedReport.java
@@ -2,6 +2,8 @@ package ic3.analyticsops.restapi.reply.print;
 
 import org.jetbrains.annotations.Nullable;
 
+import java.io.PrintStream;
+
 public class AORestApiPrintedReport
 {
     public boolean done;
@@ -11,4 +13,9 @@ public class AORestApiPrintedReport
 
     @Nullable
     public byte[] pdfContent;
+
+    public void prettyPrint(PrintStream out)
+    {
+        out.println(done + " : " + pdfName + " : " + (pdfContent != null ? pdfContent.length : 0));
+    }
 }