+++ src/main/java/ic3/analyticsops/restapi/reply/schema/AORestApiSchemaLifeCycle.java
@@ -2,6 +2,7 @@ package ic3.analyticsops.restapi.reply.schema;
 
 import org.jetbrains.annotations.Nullable;
 
+import java.io.PrintStream;
 import java.util.List;
 
 public class AORestApiSchemaLifeCycle
@@ -18,4 +19,9 @@ public class AORestApiSchemaLifeCycle
 
     @Nullable
     public List<String> errors;
+
+    public void prettyPrint(PrintStream out)
+    {
+        out.println(schemaName + " : " + status);
+    }
 }