+++ src/main/java/ic3/analyticsops/test/AOTaskContext.java
@@ -1,24 +1,28 @@
 package ic3.analyticsops.test;
 
+import ic3.analyticsops.common.AOLoggers;
 import ic3.analyticsops.restapi.client.AORestApiClientOptions;
 import ic3.analyticsops.restapi.error.AORestApiException;
 import ic3.analyticsops.restapi.request.AORestApiRequest;
 import ic3.analyticsops.test.task.reporting.AOChromeException;
 import io.webfolder.cdp.session.Session;
-import org.jetbrains.annotations.Nullable;
 
+import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.PrintStream;
+import java.lang.reflect.Method;
+import java.nio.charset.StandardCharsets;
 
 public class AOTaskContext
 {
     private final AOActorContext context;
 
-    private final AORestApiClientOptions options;
+    private final AOTask<?> task;
 
-    public AOTaskContext(AOActorContext context, @Nullable AORestApiClientOptions options)
+    public AOTaskContext(AOActorContext context, AOTask<?> task)
     {
         this.context = context;
-        this.options = options;
+        this.task = task;
     }
 
     public String getRestApiURL()
@@ -64,7 +68,10 @@ public class AOTaskContext
     public <REPLY> REPLY sendRequest(AORestApiRequest<REPLY> request)
             throws AORestApiException
     {
-        return context.sendRequest(request, options);
+        final AORestApiClientOptions options = new AORestApiClientOptions()
+                .dumpJson(task.isDumpJson());
+
+        return prettyPrint(context.sendRequest(request, options));
     }
 
     /**
@@ -73,10 +80,39 @@ public class AOTaskContext
     public <REPLY> REPLY sendRequestWithJson(AORestApiRequest<REPLY> request)
             throws AORestApiException
     {
-        return context.sendRequest(request, new AORestApiClientOptions()
-                .dumpJson(options != null && options.dumpJson)
-                .withJson(true)
-        );
+        final AORestApiClientOptions options = new AORestApiClientOptions()
+                .dumpJson(task.isDumpJson())
+                .withJson(true);
+
+        return prettyPrint(context.sendRequest(request, options));
+    }
+
+    protected <REPLY> REPLY prettyPrint(REPLY reply)
+    {
+        if (!task.isDumpResult())
+        {
+            return reply;
+        }
+
+        try
+        {
+            // NO base class for the reply payload :-(
+
+            final Method prettyPrint = reply.getClass().getMethod("prettyPrint", PrintStream.class);
+
+            final ByteArrayOutputStream out = new ByteArrayOutputStream();
+            final PrintStream ps = new PrintStream(out, true, StandardCharsets.UTF_8);
+
+            prettyPrint.invoke(reply, ps);
+
+            AOLoggers.PRETTY_PRINT.warn("[pretty-print] {}\n{}", reply.getClass().getSimpleName(), out.toString(StandardCharsets.UTF_8));
+        }
+        catch (Exception ex)
+        {
+            AOLoggers.PRETTY_PRINT.warn("[pretty-print] pretty-print missing for class : {}", reply.getClass().getSimpleName());
+        }
+
+        return reply;
     }
 
     public boolean isOnError()